refactor: promote markdown-to-html to shared utils
